requirements.txt
streamlit>=1.25.0
langchain>=0.1.0
pypdf>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
tiktoken>=0.5.1
huggingface-hub>=0.18.0